move markup related files to proper folder
