+++ src/bug-filing/azure-boards/create-azure-boards-issue-filing-url.ts
@@ -4,10 +4,10 @@ import { EnvironmentInfo } from '../../common/environment-info-provider';
 import { CreateIssueDetailsTextData } from '../../common/types/create-issue-details-text-data';
 import { title } from '../../content/strings/application';
 import { createIssueDetailsBuilder } from '../common/create-issue-details-builder';
-import { HTMLFactory } from '../common/html-factory';
 import { HTTPQueryBuilder } from '../common/http-query-builder';
 import { IssueDetailsBuilder } from '../common/issue-details-builder';
 import { IssueFilingUrlStringUtils, IssueUrlCreationUtils } from '../common/issue-filing-url-string-utils';
+import { HTMLFactory } from '../common/markup/html-factory';
 import { AzureBoardsBugFilingSettings } from './azure-boards-bug-filing-service';
 
 const buildTags = (createBugData: CreateIssueDetailsTextData, standardTags: string[]): string => {
