+++ src/bug-filing/common/markup/markdown-factory.ts
